CU-24wgdcy controllerにフォルダを追加して、ディレクトリ構成を変更した
