+++ backend/app/Http/Controllers/Blogs/BlogGoodController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Blogs;
 
-use App\Http\Requests\StoreCountryRequest;
-use App\Http\Requests\UpdateCountryRequest;
-use App\Models\Country;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreBlogGoodRequest;
+use App\Http\Requests\UpdateBlogGoodRequest;
+use App\Models\BlogGood;
 
-class CountryController extends Controller
+class BlogGoodController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -31,10 +32,10 @@ class CountryController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\StoreCountryRequest  $request
+     * @param  \App\Http\Requests\StoreBlogGoodRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreCountryRequest $request)
+    public function store(StoreBlogGoodRequest $request)
     {
         //
     }
@@ -42,10 +43,10 @@ class CountryController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Country  $country
+     * @param  \App\Models\BlogGood  $blogGood
      * @return \Illuminate\Http\Response
      */
-    public function show(Country $country)
+    public function show(BlogGood $blogGood)
     {
         //
     }
@@ -53,10 +54,10 @@ class CountryController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Country  $country
+     * @param  \App\Models\BlogGood  $blogGood
      * @return \Illuminate\Http\Response
      */
-    public function edit(Country $country)
+    public function edit(BlogGood $blogGood)
     {
         //
     }
@@ -64,11 +65,11 @@ class CountryController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\UpdateCountryRequest  $request
-     * @param  \App\Models\Country  $country
+     * @param  \App\Http\Requests\UpdateBlogGoodRequest  $request
+     * @param  \App\Models\BlogGood  $blogGood
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateCountryRequest $request, Country $country)
+    public function update(UpdateBlogGoodRequest $request, BlogGood $blogGood)
     {
         //
     }
@@ -76,10 +77,10 @@ class CountryController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Country  $country
+     * @param  \App\Models\BlogGood  $blogGood
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Country $country)
+    public function destroy(BlogGood $blogGood)
     {
         //
     }
