+++ backend/app/Http/Controllers/Blogs/BlogResponseController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Blogs;
 
-use App\Http\Requests\StoreMemoResponseGoodRequest;
-use App\Http\Requests\UpdateMemoResponseGoodRequest;
-use App\Models\MemoResponseGood;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreBlogResponseRequest;
+use App\Http\Requests\UpdateBlogResponseRequest;
+use App\Models\BlogResponse;
 
-class MemoResponseGoodController extends Controller
+class BlogResponseController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -31,10 +32,10 @@ class MemoResponseGoodController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\StoreMemoResponseGoodRequest  $request
+     * @param  \App\Http\Requests\StoreBlogResponseRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreMemoResponseGoodRequest $request)
+    public function store(StoreBlogResponseRequest $request)
     {
         //
     }
@@ -42,10 +43,10 @@ class MemoResponseGoodController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\MemoResponseGood  $memoResponseGood
+     * @param  \App\Models\BlogResponse  $blogResponse
      * @return \Illuminate\Http\Response
      */
-    public function show(MemoResponseGood $memoResponseGood)
+    public function show(BlogResponse $blogResponse)
     {
         //
     }
@@ -53,10 +54,10 @@ class MemoResponseGoodController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\MemoResponseGood  $memoResponseGood
+     * @param  \App\Models\BlogResponse  $blogResponse
      * @return \Illuminate\Http\Response
      */
-    public function edit(MemoResponseGood $memoResponseGood)
+    public function edit(BlogResponse $blogResponse)
     {
         //
     }
@@ -64,11 +65,11 @@ class MemoResponseGoodController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\UpdateMemoResponseGoodRequest  $request
-     * @param  \App\Models\MemoResponseGood  $memoResponseGood
+     * @param  \App\Http\Requests\UpdateBlogResponseRequest  $request
+     * @param  \App\Models\BlogResponse  $blogResponse
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateMemoResponseGoodRequest $request, MemoResponseGood $memoResponseGood)
+    public function update(UpdateBlogResponseRequest $request, BlogResponse $blogResponse)
     {
         //
     }
@@ -76,10 +77,10 @@ class MemoResponseGoodController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\MemoResponseGood  $memoResponseGood
+     * @param  \App\Models\BlogResponse  $blogResponse
      * @return \Illuminate\Http\Response
      */
-    public function destroy(MemoResponseGood $memoResponseGood)
+    public function destroy(BlogResponse $blogResponse)
     {
         //
     }
