--- backend/app/Http/Controllers/FriendController.php
@@ -1,86 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Http\Requests\StoreFriendRequest;
-use App\Http\Requests\UpdateFriendRequest;
-use App\Models\Friend;
-
-class FriendController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \App\Http\Requests\StoreFriendRequest  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(StoreFriendRequest $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Friend  $friend
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Friend $friend)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Friend  $friend
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Friend $friend)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \App\Http\Requests\UpdateFriendRequest  $request
-     * @param  \App\Models\Friend  $friend
-     * @return \Illuminate\Http\Response
-     */
-    public function update(UpdateFriendRequest $request, Friend $friend)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Friend  $friend
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Friend $friend)
-    {
-        //
-    }
-}
