+++ backend/app/Http/Controllers/Hitchhikes/Doings/DriverReviewController.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Hitchhikes\Doings;
 
-use App\Http\Requests\StoreBlogRequest;
-use App\Http\Requests\UpdateBlogRequest;
-use App\Models\Blog;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreDriverReviewRequest;
+use App\Http\Requests\UpdateDriverReviewRequest;
+use App\Models\DriverReview;
 
-class BlogController extends Controller
+class DriverReviewController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -31,10 +32,10 @@ class BlogController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\StoreBlogRequest  $request
+     * @param  \App\Http\Requests\StoreDriverReviewRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreBlogRequest $request)
+    public function store(StoreDriverReviewRequest $request)
     {
         //
     }
@@ -42,10 +43,10 @@ class BlogController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Blog  $blog
+     * @param  \App\Models\DriverReview  $driverReview
      * @return \Illuminate\Http\Response
      */
-    public function show(Blog $blog)
+    public function show(DriverReview $driverReview)
     {
         //
     }
@@ -53,10 +54,10 @@ class BlogController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Blog  $blog
+     * @param  \App\Models\DriverReview  $driverReview
      * @return \Illuminate\Http\Response
      */
-    public function edit(Blog $blog)
+    public function edit(DriverReview $driverReview)
     {
         //
     }
@@ -64,11 +65,11 @@ class BlogController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\UpdateBlogRequest  $request
-     * @param  \App\Models\Blog  $blog
+     * @param  \App\Http\Requests\UpdateDriverReviewRequest  $request
+     * @param  \App\Models\DriverReview  $driverReview
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateBlogRequest $request, Blog $blog)
+    public function update(UpdateDriverReviewRequest $request, DriverReview $driverReview)
     {
         //
     }
@@ -76,10 +77,10 @@ class BlogController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Blog  $blog
+     * @param  \App\Models\DriverReview  $driverReview
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Blog $blog)
+    public function destroy(DriverReview $driverReview)
     {
         //
     }
