+++ backend/app/Http/Controllers/Hitchhikes/Doings/HitchhikeMessageController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Hitchhikes\Doings;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreHitchhikeMessageRequest;
 use App\Http\Requests\UpdateHitchhikeMessageRequest;
 use App\Models\HitchhikeMessage;
