+++ backend/app/Http/Controllers/Hitchhikes/Spots/HitchhikeResponseController.php
@@ -1,0 +1,87 @@
+<?php
+
+namespace App\Http\Controllers\Hitchhikes\Spots;
+
+use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreHitchhikeResponseRequest;
+use App\Http\Requests\UpdateHitchhikeResponseRequest;
+use App\Models\HitchhikeResponse;
+
+class HitchhikeResponseController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \App\Http\Requests\StoreHitchhikeResponseRequest  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(StoreHitchhikeResponseRequest $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\HitchhikeResponse  $hitchhikeResponse
+     * @return \Illuminate\Http\Response
+     */
+    public function show(HitchhikeResponse $hitchhikeResponse)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\Models\HitchhikeResponse  $hitchhikeResponse
+     * @return \Illuminate\Http\Response
+     */
+    public function edit(HitchhikeResponse $hitchhikeResponse)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \App\Http\Requests\UpdateHitchhikeResponseRequest  $request
+     * @param  \App\Models\HitchhikeResponse  $hitchhikeResponse
+     * @return \Illuminate\Http\Response
+     */
+    public function update(UpdateHitchhikeResponseRequest $request, HitchhikeResponse $hitchhikeResponse)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Models\HitchhikeResponse  $hitchhikeResponse
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy(HitchhikeResponse $hitchhikeResponse)
+    {
+        //
+    }
+}
